refactor: using controller at resolvers
