+++ infrastructure/graph/resolver.go
@@ -1,6 +1,9 @@
 package graph
 
-import "github.com/clock-en/go-todo-on-ddd-on-ddd/infrastructure/graph/model"
+import (
+	"github.com/clock-en/go-todo-on-ddd-on-ddd/infrastructure/graph/generated"
+	"github.com/clock-en/go-todo-on-ddd-on-ddd/infrastructure/graph/model"
+)
 
 // This file will not be regenerated automatically.
 //
@@ -9,3 +12,12 @@ import "github.com/clock-en/go-todo-on-ddd-on-ddd/infrastructure/graph/model"
 type Resolver struct {
 	tasks []*model.Task
 }
+
+// Mutation returns generated.MutationResolver implementation.
+func (r *Resolver) Mutation() generated.MutationResolver { return &mutationResolver{r} }
+
+// Query returns generated.QueryResolver implementation.
+func (r *Resolver) Query() generated.QueryResolver { return &queryResolver{r} }
+
+type mutationResolver struct{ *Resolver }
+type queryResolver struct{ *Resolver }
